refactor(vm): Move api macros to a separate module
Reduces the size of the api module and lets the serde module use the
macros in tests
